refactor: split file preview native sessions
