+++ Sources/Panels/FilePreviewPanel.swift
@@ -1185,387 +1185,6 @@ final class FilePreviewPanel: Panel, ObservableObject, FilePreviewTextEditingPan
     }
 }
 
-@MainActor
-final class PanelOwnedNativeViewSession<View: NSView> {
-    private let makeView: @MainActor () -> View
-    private let closeView: @MainActor (View) -> Void
-    private var ownedView: View?
-
-    init(
-        makeView: @escaping @MainActor () -> View,
-        closeView: @escaping @MainActor (View) -> Void = { $0.removeFromSuperview() }
-    ) {
-        self.makeView = makeView
-        self.closeView = closeView
-    }
-
-    func view(configure: @MainActor (View) -> Void) -> View {
-        let view = ownedView ?? makeView()
-        ownedView = view
-        if view.superview != nil {
-            view.removeFromSuperview()
-        }
-        configure(view)
-        return view
-    }
-
-    func update(_ view: View, configure: @MainActor (View) -> Void) {
-        if ownedView == nil {
-            ownedView = view
-        }
-        configure(view)
-    }
-
-    func close() {
-        if let ownedView {
-            closeView(ownedView)
-        }
-        ownedView = nil
-    }
-}
-
-@MainActor
-final class FilePreviewNativeViewSessions {
-    let pdf = FilePreviewPDFSession()
-    let image = FilePreviewImageSession()
-    let media = FilePreviewMediaSession()
-    let quickLook = FilePreviewQuickLookSession()
-
-    func closeInactive(except mode: FilePreviewMode) {
-        switch mode {
-        case .text:
-            closeAll()
-        case .pdf:
-            image.close()
-            media.close()
-            quickLook.close()
-        case .image:
-            pdf.close()
-            media.close()
-            quickLook.close()
-        case .media:
-            pdf.close()
-            image.close()
-            quickLook.close()
-        case .quickLook:
-            pdf.close()
-            image.close()
-            media.close()
-        }
-    }
-
-    func closeAll() {
-        pdf.close()
-        image.close()
-        media.close()
-        quickLook.close()
-    }
-}
-
-@MainActor
-final class FilePreviewPDFSession {
-    private let viewSession = PanelOwnedNativeViewSession(
-        makeView: FilePreviewPDFContainerView.init,
-        closeView: { $0.close() }
-    )
-
-    func view(
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) -> FilePreviewPDFContainerView {
-        viewSession.view {
-            configure(
-                $0,
-                panel: panel,
-                isVisibleInUI: isVisibleInUI,
-                backgroundColor: backgroundColor,
-                drawsBackground: drawsBackground
-            )
-        }
-    }
-
-    func update(
-        _ view: FilePreviewPDFContainerView,
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) {
-        viewSession.update(view) {
-            configure(
-                $0,
-                panel: panel,
-                isVisibleInUI: isVisibleInUI,
-                backgroundColor: backgroundColor,
-                drawsBackground: drawsBackground
-            )
-        }
-    }
-
-    func close() {
-        viewSession.close()
-    }
-
-    private func configure(
-        _ view: FilePreviewPDFContainerView,
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) {
-        view.isHidden = !isVisibleInUI
-        view.setBackgroundAppearance(backgroundColor: backgroundColor, drawsBackground: drawsBackground)
-        view.setPanel(panel)
-        view.setURL(panel.fileURL)
-    }
-}
-
-@MainActor
-final class FilePreviewImageSession {
-    private let viewSession = PanelOwnedNativeViewSession(
-        makeView: FilePreviewImageContainerView.init,
-        closeView: { $0.close() }
-    )
-
-    func view(
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) -> FilePreviewImageContainerView {
-        viewSession.view {
-            configure(
-                $0,
-                panel: panel,
-                isVisibleInUI: isVisibleInUI,
-                backgroundColor: backgroundColor,
-                drawsBackground: drawsBackground
-            )
-        }
-    }
-
-    func update(
-        _ view: FilePreviewImageContainerView,
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) {
-        viewSession.update(view) {
-            configure(
-                $0,
-                panel: panel,
-                isVisibleInUI: isVisibleInUI,
-                backgroundColor: backgroundColor,
-                drawsBackground: drawsBackground
-            )
-        }
-    }
-
-    func close() {
-        viewSession.close()
-    }
-
-    private func configure(
-        _ view: FilePreviewImageContainerView,
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) {
-        view.isHidden = !isVisibleInUI
-        view.setBackgroundAppearance(backgroundColor: backgroundColor, drawsBackground: drawsBackground)
-        view.setPanel(panel)
-        view.setURL(panel.fileURL)
-    }
-}
-
-@MainActor
-final class FilePreviewMediaSession {
-    private let viewSession = PanelOwnedNativeViewSession<AVPlayerView>(
-        makeView: FilePreviewMediaSession.makeView,
-        closeView: { view in
-            view.player = nil
-            view.removeFromSuperview()
-        }
-    )
-    private var currentURL: URL?
-    private var player: AVPlayer?
-
-    deinit {
-        player?.pause()
-    }
-
-    func view(
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) -> AVPlayerView {
-        viewSession.view {
-            configure(
-                $0,
-                panel: panel,
-                isVisibleInUI: isVisibleInUI,
-                backgroundColor: backgroundColor,
-                drawsBackground: drawsBackground
-            )
-        }
-    }
-
-    func update(
-        _ view: AVPlayerView,
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) {
-        viewSession.update(view) {
-            configure(
-                $0,
-                panel: panel,
-                isVisibleInUI: isVisibleInUI,
-                backgroundColor: backgroundColor,
-                drawsBackground: drawsBackground
-            )
-        }
-    }
-
-    func close() {
-        player?.pause()
-        viewSession.close()
-        player = nil
-        currentURL = nil
-    }
-
-    private static func makeView() -> AVPlayerView {
-        let view = AVPlayerView()
-        view.controlsStyle = .floating
-        view.showsFullScreenToggleButton = true
-        view.videoGravity = .resizeAspect
-        return view
-    }
-
-    private func configure(
-        _ view: AVPlayerView,
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) {
-        view.isHidden = !isVisibleInUI
-        FilePreviewNativeBackground.applyRootLayer(
-            to: view,
-            backgroundColor: backgroundColor,
-            drawsBackground: drawsBackground
-        )
-        panel.attachPreviewFocus(root: view, primaryResponder: view, intent: .mediaPlayer)
-        updatePlayer(in: view, url: panel.fileURL)
-    }
-
-    private func updatePlayer(in playerView: AVPlayerView, url: URL) {
-        guard currentURL != url else { return }
-        player?.pause()
-        currentURL = url
-        let player = AVPlayer(url: url)
-        self.player = player
-        playerView.player = player
-    }
-}
-
-@MainActor
-final class FilePreviewQuickLookSession {
-    private let viewSession = PanelOwnedNativeViewSession<NSView>(
-        makeView: FilePreviewQuickLookSession.makeView,
-        closeView: { view in
-            if let previewView = view as? QLPreviewView {
-                previewView.close()
-                previewView.previewItem = nil
-            }
-            view.removeFromSuperview()
-        }
-    )
-    private var item: FilePreviewQLItem?
-
-    func view(
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) -> NSView {
-        viewSession.view {
-            configure(
-                $0,
-                panel: panel,
-                isVisibleInUI: isVisibleInUI,
-                backgroundColor: backgroundColor,
-                drawsBackground: drawsBackground
-            )
-        }
-    }
-
-    func update(
-        _ view: NSView,
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) {
-        viewSession.update(view) {
-            configure(
-                $0,
-                panel: panel,
-                isVisibleInUI: isVisibleInUI,
-                backgroundColor: backgroundColor,
-                drawsBackground: drawsBackground
-            )
-        }
-    }
-
-    func close() {
-        viewSession.close()
-        item = nil
-    }
-
-    private static func makeView() -> NSView {
-        guard let previewView = QLPreviewView(frame: .zero, style: .normal) else {
-            return NSView()
-        }
-        previewView.autostarts = true
-        return previewView
-    }
-
-    private func configure(
-        _ view: NSView,
-        panel: FilePreviewPanel,
-        isVisibleInUI: Bool,
-        backgroundColor: NSColor,
-        drawsBackground: Bool
-    ) {
-        view.isHidden = !isVisibleInUI
-        if let previewView = view as? QLPreviewView {
-            panel.attachPreviewFocus(root: previewView, primaryResponder: previewView, intent: .quickLook)
-            previewView.previewItem = previewItem(for: panel.fileURL, title: panel.displayTitle)
-        }
-        FilePreviewNativeBackground.applyRootLayer(
-            to: view,
-            backgroundColor: backgroundColor,
-            drawsBackground: drawsBackground
-        )
-    }
-
-    private func previewItem(for url: URL, title: String) -> FilePreviewQLItem {
-        if let item, item.url == url, item.title == title {
-            return item
-        }
-        let next = FilePreviewQLItem(url: url, title: title)
-        item = next
-        return next
-    }
-}
-
 struct FilePreviewPanelView: View {
     @ObservedObject var panel: FilePreviewPanel
     let isFocused: Bool
@@ -4680,24 +4299,6 @@ private struct QuickLookPreviewView: NSViewRepresentable {
             backgroundColor: backgroundColor,
             drawsBackground: drawsBackground
         )
-    }
-}
-
-private final class FilePreviewQLItem: NSObject, QLPreviewItem {
-    let url: URL
-    let title: String
-
-    init(url: URL, title: String) {
-        self.url = url
-        self.title = title
-    }
-
-    var previewItemURL: URL? {
-        url
-    }
-
-    var previewItemTitle: String? {
-        title
     }
 }
